PP-14728 - moved GatewayStatusComparison.java to model
