+++ src/main/java/uk/gov/pay/connector/paymentprocessor/model/GatewayStatusComparison.java
@@ -1,4 +1,4 @@
-package uk.gov.pay.connector.paymentprocessor.service;
+package uk.gov.pay.connector.paymentprocessor.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
